Android SV config: need to restore to device default for disable
For devices that does not yet support QMI Loc API for
constellation disablement, we will use QMI command to restore
SV type to device default for the case that subsequent request
disables a larger set of constellations than the previous request.
Without this, the constellation may get blacklisted and not disabled,
which is sub-optimal.

CRs-fixed: 3870673

Change-Id: Icbe14e5dd2cfc6197ffb18c20d869ae5195971ee